Redesigned the interface-related modules once more
For more details about and reasons for this redesign, see #54.
