+++ gdpc/minecraft_tools.py
@@ -13,7 +13,7 @@ from .vector_tools import Rect
 from .block import Block
 from . import lookup
 from .lookup import SUPPORTS, VERSIONS
-from .worldSlice import WorldSlice
+from .world_slice import WorldSlice
 
 
 def closest_version(version):
@@ -26,39 +26,40 @@ def closest_version(version):
     return 0
 
 
-def check_version():
-    """Retrieve Minecraft version and check compatibility."""
-    wslice = WorldSlice(Rect(size=ivec2(1,1))) # single-chunk slice
-    current = int(wslice.nbtfile["Chunks"][0]["DataVersion"].value)
-    closestname = "Unknown"
-    # check compatibility
-    if current not in VERSIONS or VERSIONS[SUPPORTS] not in VERSIONS[current]:
-        closest = closest_version(current)
-        closestname = VERSIONS[closest]
-        closestname += " snapshot" if current > closest else ""
-        if closest > SUPPORTS:
-            print(colored(color="yellow", text=\
-                f"WARNING: You are using a newer "
-                "version of Minecraft then GDPC supports!\n"
-                f"\tSupports: {VERSIONS[SUPPORTS]} "
-                f"Detected: {closestname}"
-            ))
-        elif closest < SUPPORTS:
-            print(colored(color="yellow", text=\
-                f"WARNING: You are using an older "
-                "version of Minecraft then GDPC supports!\n"
-                f"\tSupports: {VERSIONS[SUPPORTS]} "
-                f"Detected: {closestname}"
-            ))
-        else:
-            raise ValueError(colored(color="red", text=\
-                f"Invalid supported version: "
-                f"SUPPORTS = {current}!"
-            ))
-    else:
-        closestname = VERSIONS[current]
+# TODO: re-implement using version endpoint
+# def check_version():
+#     """Retrieve Minecraft version and check compatibility."""
+#     wslice = WorldSlice(Rect(size=ivec2(1,1))) # single-chunk slice
+#     current = int(wslice.nbtfile["Chunks"][0]["DataVersion"].value)
+#     closestname = "Unknown"
+#     # check compatibility
+#     if current not in VERSIONS or VERSIONS[SUPPORTS] not in VERSIONS[current]:
+#         closest = closest_version(current)
+#         closestname = VERSIONS[closest]
+#         closestname += " snapshot" if current > closest else ""
+#         if closest > SUPPORTS:
+#             print(colored(color="yellow", text=\
+#                 f"WARNING: You are using a newer "
+#                 "version of Minecraft then GDPC supports!\n"
+#                 f"\tSupports: {VERSIONS[SUPPORTS]} "
+#                 f"Detected: {closestname}"
+#             ))
+#         elif closest < SUPPORTS:
+#             print(colored(color="yellow", text=\
+#                 f"WARNING: You are using an older "
+#                 "version of Minecraft then GDPC supports!\n"
+#                 f"\tSupports: {VERSIONS[SUPPORTS]} "
+#                 f"Detected: {closestname}"
+#             ))
+#         else:
+#             raise ValueError(colored(color="red", text=\
+#                 f"Invalid supported version: "
+#                 f"SUPPORTS = {current}!"
+#             ))
+#     else:
+#         closestname = VERSIONS[current]
 
-    return (current, closestname)
+#     return (current, closestname)
 
 
 def positionToInventoryIndex(position: ivec2, inventorySize: ivec2):
